research_backups/requirements.txt
numpy==1.23.4
matplotlib==3.6.1 
python-binance==1.0.1
pyzmq==24.0.1 
pandas==1.5.1
scipy==1.9.3